started implementing a GP wrapper for expensive functions
Replace an expensive black-box function with a cheaper-to-evaluate
model. Each time the wrapper is called, it decides whether it’s
confident enough in the predictions to yield the predicted output or
whether it needs to call the expensive ground-truth function. Whenever
it calls the true function, it updates the model.